+++ .travis.yml
@@ -7,6 +7,7 @@ install:
   - pip install autodocsumm sphinx_rtd_theme
 
 script:
+  - PYTHONHASHSEED=32 pytest
   - python setup.py sdist
   - python setup.py bdist_wheel -p any
   - rm -rf build
